% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CalciumSpikes.R
\docType{package}
\name{CalciumSpikes}
\alias{CalciumSpikes}
\title{CalciumSpikes: A package for statistically modelling time-dependent calcium oscillations.}
\description{
The CalciumSpikes package provides functions relating to the statistical modeling of calcium spike sequences. This includes simulating surrogate spike sequences. Fitting models to data and assessing fitted models.
}
