% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GaussianProcess.R
\name{cred_interval_gp}
\alias{cred_interval_gp}
\title{Get mean + 95\% credible of intensity function}
\usage{
cred_interval_gp(intensities, end.time)
}
\arguments{
\item{intensities}{matrix of posterior intensity functions}

\item{end.time}{The length of the experiment}
}
\value{
either plot of mean and 95\% credible region. Or a list containing the mean and 95\% credible region of the intensity function
}
\description{
Get mean + 95\% credible of intensity function
}
