% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PWC.R
\name{cred_interval_pwc}
\alias{cred_interval_pwc}
\title{get the mean and 95\% credible interval of the intensity function for PWC method}
\usage{
cred_interval_pwc(partitions, heights, method = "True", steps = 1000)
}
\arguments{
\item{partitions}{Matrix of posteior samples of the partition of the change points}

\item{heights}{Matrix of posterior heights corresponding to the PWC functions.}

\item{method}{Either "True" or "Approx". Where Approx only calculate the intensity values at the specified steps. True calculates at every changepoint in the posterior distribution.}

\item{steps}{The number of steps to partition the intensity function into.}
}
\value{
A list containing the mean and 95\% credible interval for the intensity function.
}
\description{
get the mean and 95\% credible interval of the intensity function for PWC method
}
