% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GaussianProcess.R
\name{draw_conditionedMVN}
\alias{draw_conditionedMVN}
\title{Draw from a conditioned MVN}
\usage{
draw_conditionedMVN(
  times,
  x,
  sigma.var,
  l,
  pos = "start",
  noise = 0.001,
  do.log = T,
  known = 10,
  f = "current"
)
}
\arguments{
\item{times}{va}

\item{x}{va}

\item{sigma.var}{va}

\item{l}{va}

\item{pos}{va}

\item{noise}{va}

\item{do.log}{va}

\item{known}{va}

\item{f}{va}
}
\description{
Draw from a conditioned MVN
}
