% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QQ_KS.R
\name{get.tau}
\alias{get.tau}
\title{A function that returns the rescaled ISI times.}
\usage{
get.tau(R.t, spike.times, end.time)
}
\arguments{
\item{R.t}{R.t}

\item{spike.times}{spike.times}

\item{end.time}{end.time}
}
\value{
rescaled ISI
}
\description{
A function that returns the rescaled ISI times.
}
