% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QQ_KS.R
\name{get_slopes}
\alias{get_slopes}
\title{A function that returns the slopes of the individual spike cell data of QQ and KS plots.}
\usage{
get_slopes(data, spikes)
}
\arguments{
\item{data}{The QQ and KS quantiles obtained from \code{QQ_KS_data}}

\item{spikes}{The calcium spike sequences used to fit the model (a data frame where each column consists of a single spike sequence).}
}
\value{
slopes of the QQ and KS plots.
}
\description{
A function that returns the slopes of the individual spike cell data of QQ and KS plots.
}
