% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GaussianProcess.R
\name{log_pi_l_chol}
\alias{log_pi_l_chol}
\title{log_pi_l with cholesky decomposition}
\usage{
log_pi_l_chol(l, x, cov_chol)
}
\arguments{
\item{l}{Length scale}

\item{x}{Intensity function}

\item{cov_chol}{the cholesky decomposition of the covariance matrix}
}
\value{
Marginal of the length scale
}
\description{
log_pi_l with cholesky decomposition
}
