% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Constant.R
\name{mcmc_constant}
\alias{mcmc_constant}
\title{Model fitting when the intensity function is assumed to be constant.}
\usage{
mcmc_constant(
  d.spikes,
  end.time,
  prior.hyper = c(1, 0.01),
  prior.x = c(1, 0.01),
  prior.tmin = c(1, 0.01),
  sigma.hyper,
  sigma.x,
  sigma.t.min,
  iter,
  burn,
  ISI.type,
  show.prog = T
)
}
\arguments{
\item{d.spikes}{The calcium spike sequences used to fit the model (a data frame where each column consists of a single spike sequence).}

\item{end.time}{The length of the experiment}

\item{prior.hyper}{Either a fixed value for the ISI parameter or parameters for the gamma prior for the ISI parameter..}

\item{prior.x}{Either a fixed value for the intensity function or parameters for the gamma prior for the intensity function.}

\item{prior.tmin}{prior.tmin}

\item{sigma.hyper}{RW-metropolis variance for the ISI parameter.}

\item{sigma.x}{RW-metropolis variance for the intensity function.}

\item{sigma.t.min}{RW-metropolis variance for the refractory period.}

\item{iter}{The number of iterations of the MCMC after the initial burn in.}

\item{burn}{The initial number of iterations to remove.}

\item{ISI.type}{The inhomogeneous ISI distribution. Either 'Exponential', 'Gamma', 'InverseGaussian', 'LogNormal' or 'Weibull'.}

\item{show.prog}{Boolean flag for do we print a progress bar of the samples.}
}
\value{
\code{iter} samples of the posterior ISI distribution.
}
\description{
This function runs an MCMC algorithm to sample from the posterior ISI distribution given calcium spike sequences, where we assume the intensity function is constant.
}
\examples{
\dontrun{
spikes <- sort(runif(20, min=0,max=20))
posterior = mcmc_constant(data.frame(spikes), end.time=20, iter = 10000,
burn = 5000, ISI.type='Gamma', sigma.x=1, sigma.hyper = 1,prior.tmin=0)
}
}
