% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PWC.R
\name{mcmc_pwc}
\alias{mcmc_pwc}
\title{Model fitting when the intensity function is assumed to have a PWC prior.}
\usage{
mcmc_pwc(
  d.spikes,
  end.time,
  iter,
  burn,
  k.max,
  lambda,
  kappa,
  mu,
  kappa0,
  hyper.param = c(1, 0.001),
  sigma.h = NULL,
  start.hyper = 1000,
  hyper.initial = 0.1,
  T.min.param = NULL,
  T.min.initial = 0.05,
  sigma.t = NULL,
  ISI.type = "Gamma",
  do.log = TRUE,
  show.iter = T,
  which.heights = "independent"
)
}
\arguments{
\item{d.spikes}{The calcium spike sequences used to fit the model (a data frame where each column consists of a single spike sequence).}

\item{end.time}{The length of the experiment}

\item{iter}{The number of iterations of the MCMC after the initial burn in.}

\item{burn}{The initial number of iterations to remove.}

\item{k.max}{Maximum number of change points in RJMCMC.}

\item{lambda}{Parameter of the Poisson distribution for the number of change points.}

\item{kappa}{parameter of priors height}

\item{mu}{parameter of priors height}

\item{kappa0}{parameter of priors height}

\item{hyper.param}{Either the fixed value of the ISI parameter or the parameters of the Gamma prior for the ISI parameter.}

\item{sigma.h}{RW-metropolis variance for the ISI parameter.}

\item{start.hyper}{The iteration to begin inferring the ISI parameter.}

\item{hyper.initial}{Initial value of the ISI parameter.}

\item{T.min.param}{Either the fixed value of the refractory period or the parameters of the Gamma prior for the refractory period. If we do not want to include the refractory period set to NA.}

\item{T.min.initial}{Initial value of the refractory period.}

\item{sigma.t}{RW-metropolis variance for the refractory period.}

\item{ISI.type}{The inhomogeneous ISI distribution. Either 'Exponential', 'Gamma', 'InverseGaussian', 'LogNormal' or 'Weibull'.}

\item{do.log}{Flag, where if do.log is true the calculations are computed on the log scale.}

\item{show.iter}{Flag for whether to print the number of iterations complete into the console.}

\item{which.heights}{Method for prior distribution of heights, either 'independent' or 'martingale'.}
}
\value{
\code{iter} samples of the posterior ISI distribution
}
\description{
This function runs an MCMC algorithm to sample from the posterior ISI distribution given calcium spike sequences, where we assume the intensity function has a piecewise constant prior.
}
\examples{
\dontrun{
spikes <- sort(runif(20, min=0,max=20))
posterior = mcmc_pwc(data.frame(spikes), end.time=20, iter = 10000, burn = 5000,
 ISI.type='Gamma', sigma.h = 1, k.max=30, lambda=20, kappa=1,mu=1,kappa0=1)
}
}
