% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QQ_KS.R
\name{QQ_KS_data}
\alias{QQ_KS_data}
\title{Get the QQ and KS quantiles}
\usage{
QQ_KS_data(
  spikes.df,
  int.fn,
  end.time,
  ISI.type,
  hyper.param,
  t.min = 0,
  rm.last.spike = TRUE,
  do.log = TRUE
)
}
\arguments{
\item{spikes.df}{The calcium spike sequences used to fit the model (a data frame where each column consists of a single spike sequence).}

\item{int.fn}{The intensity function}

\item{end.time}{The length of the experiment}

\item{ISI.type}{The inhomogeneous ISI distribution. Either 'Exponential', 'Gamma', 'InverseGaussian', 'LogNormal' or 'Weibull'.}

\item{hyper.param}{The ISI parameter}

\item{t.min}{the refractory period}

\item{rm.last.spike}{Flag for do we remove the last spike time (it may correspond to the end.time)}

\item{do.log}{Flag for do we calculate on the logarithmic scale}
}
\value{
QQ and KS quantiles
}
\description{
Get the QQ and KS quantiles
}
