% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QQ_KS.R
\name{R.t}
\alias{R.t}
\title{A function that returns R(t), the integrated conditional intensity function.}
\usage{
R.t(q, time.step)
}
\arguments{
\item{q}{the conditional intensity function}

\item{time.step}{time.step of the conditional intensity function.}
}
\value{
the integrated intensity function
}
\description{
A function that returns R(t), the integrated conditional intensity function.
}
