% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runApps.R
\name{run_app}
\alias{run_app}
\title{Applications for Simulating Spikes}
\usage{
run_app(app)
}
\arguments{
\item{app}{The type of application you want to open. The options are 'SimulatingSpikes' or 'ViewData'}
}
\description{
Using run_app() opens either the 'ViewData' or 'SimulatingSpikes' applications.
}
\section{'ViewData'}{

An app that allows raw time-series calcium concentration data to be thresholded to obtain spike sequences.
}

\section{'SimulatingSpikes'}{

An app that allows the user to select an ISI distribution and simulate spike sequences from it.
The application only allows for the implementation of 5 ISI types, namely: Exponential, Gamma, Inverse Gaussian, Log Normal and Weibull. To use alternatively distributions/parameterisations please use or \code{\link{simulate_spikes}}.
}

\examples{
\dontrun{
run_app('ViewData')
run_app('SimulatingSpikes')
}
}
